wrapper for next-auth
