+++ packages/js-sdk/src/index.js
@@ -1,0 +1,1 @@
+export { KeypAuth } from "./keyp-auth";
